# contracts/generated/internal.openapi.json
{
...
  "info": {
    "title": "AI-CRM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/application-registry": {
      "get": {
        "description": "Returns only enabled internal applications, routes, and ancestor-closed navigation currently authorized for the internal principal. Registrations for other audiences and unauthorized targets are filtered server-side.",
        "operationId": "getInternalApplicationRegistry",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "applications": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "applicationId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          },
                          "audience": {
                            "const": "internal"
                          },
                          "enabled": {
                            "const": true
                          },
                          "permissionCode": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,63}:[a-z][a-z0-9_.-]{0,63}$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "applicationId",
                          "audience",
                          "enabled",
                          "permissionCode"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "navigation": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "applicationId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          },
                          "enabled": {
                            "const": true
                          },
                          "navigationId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          },
                          "order": {
                            "maximum": 100000,
                            "minimum": 0,
                            "type": "integer"
                          },
                          "parentNavigationId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          },
                          "routeId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "applicationId",
                          "enabled",
                          "navigationId",
                          "order",
                          "routeId"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "routes": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "applicationId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          },
                          "deepLinkSources": {
                            "items": {
                              "enum": [
                                "task",
                                "notification"
                              ],
                              "type": "string"
                            },
                            "maxItems": 2,
                            "type": "array",
                            "uniqueItems": true
                          },
                          "enabled": {
                            "const": true
                          },
                          "path": {
                            "pattern": "^/[a-z0-9_./:-]{0,255}$",
                            "type": "string"
                          },
                          "permissionCode": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,63}:[a-z][a-z0-9_.-]{0,63}$",
                            "type": "string"
                          },
                          "routeId": {
                            "pattern": "^[a-z][a-z0-9_.-]{0,127}$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "applicationId",
                          "deepLinkSources",
                          "enabled",
                          "path",
                          "permissionCode",
                          "routeId"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "version": {
                      "const": 1
                    }
                  },
                  "required": [
                    "applications",
                    "navigation",
                    "routes",
                    "version"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Current authorized internal registry snapshot."
          },
          "401": {
            "description": "No valid internal BFF session."
          },
          "403": {
            "description": "Current server-side or target authorization denied."
          },
          "503": {
            "description": "Authorization or registry persistence is unavailable; fails closed."
          }
        },
        "x-ai-crm-audiences": [
          "internal"
        ],
        "x-ai-crm-csrf": {
          "mode": "not-required",
          "reason": "read-only"
        },
        "x-ai-crm-idempotency": {
          "mode": "none",
          "reason": "read-only"
        },
        "x-ai-crm-permission": {
          "action": "read",
          "code": "platform.app-registry.registry:read",
          "owner": "platform.app-registry",
          "resource": "platform.app-registry.registry",
          "version": 1
        }
      }
    },
...
  },
...
}